backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
supabase
python-dotenv
twilio
apscheduler
pytz
python-multipart
httpx
scikit-learn
xgboost
numpy
pandas
razorpay
requests
redis
